Create model TaskTrackModel & rename VTaskTrackModel
