+++ DIMS-Core/DIMS-Core/Models/TaskTrack/TaskTrackViewModel.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace DIMS_Core.Models.TaskTrack
+{
+    public class TaskTrackViewModel
+    {
+        public int TaskTrackId { get; set; }
+        public int UserTaskId { get; set; }
+        public DateTime TrackDate { get; set; }
+        public string TrackNote { get; set; }
+    }
+}
